use binding with increase and back off

+++ auction/src/template/AuctionTemplate.java
@@ -19,7 +19,7 @@ import logist.topology.Topology.City;
 /**
  * A very simple auction agent that assigns all tasks to its first vehicle and
  * handles them sequentially.
- * 
+ *
  */
 @SuppressWarnings("unused")
 public class AuctionTemplate implements AuctionBehavior {
@@ -51,7 +51,7 @@ public class AuctionTemplate implements AuctionBehavior {
 			currentCity = previous.deliveryCity;
 		}
 	}
-	
+
 	@Override
 	public Long askPrice(Task task) {
 
@@ -67,12 +67,12 @@ public class AuctionTemplate implements AuctionBehavior {
 		double ratio = 1.0 + (random.nextDouble() * 0.05 * task.id);
 		double bid = ratio * marginalCost;
 
-		return (long) Math.round(bid);
+		return Math.round(bid);
 	}
 
 	@Override
 	public List<Plan> plan(List<Vehicle> vehicles, TaskSet tasks) {
-		
+
 //		System.out.println("Agent " + agent.id() + " has tasks " + tasks);
 
 		Plan planVehicle1 = naivePlan(vehicle, tasks);
